Content fixes and tweaks from initial site upload.
This CL adds a few files that were missed in the initial
site upload and also fixes some issues caused by switching
to the "modern CSS reset" in the CSS stylesheet.

It should be fine to push the content live after this (though
obviously there's plenty of things to fix still).

Change-Id: I37d5ee572340d2c5a996c43b41d479fc31968d43