Implement support for 'listpage' and 'listitem' in exports.
Google Sites supported a mechanism for authoring tables of
structured data using a 'listpage' and 'listitems'. We
used this on a few pages, most notably on
/chromium-os/developer-information-for-chrome-os-devices.

Prior versions of the export code didn't support the listitems,
causing those tables to be left out.

This CL adds basic support for the data types, and so now we'll
export these as generic HTML tables. The actual Google Sites
tables had a nice dynamic datagrid sort of web component instead
(where you could sort by each column).

We should add support for that as well, but that'll be a separate
bug and change.

Bug: 1267382
Change-Id: I9fc1bf091474d903a076681b16d985bcb9405829