Add a 'chromium-website-staging' Firebase project.
This adds a staging project to Firebase and makes it the default.

This (and subsequent changes) will make it a bit harder to
accidentally deploy straight to prod when testing changes to
the infrastructure.

Once the infra code is working, the bots will be changed to use
prod by default.

Bug: 1268676, 1267501
Change-Id: I46368c4fa1b80a078c7e6704f6bf61bbbcf54582