Clean up /developers/about-signin-internals
This fixes the default styling of list items to have
list-style-position: outside (so that the second and
subsequent lines of a list item are indented), and
fixes the weird formatting bug that was causing everything
to be bold and/or list items to be indented.

There's probably bugs in the export to be fixed b/c of this,
but fix this page by hand for now.

Bug: 1260513
Change-Id: I2461a8877d45fc18672ad65a0e80f2433a95e4d9