Import new content and fix ':' in filenames.
Some of the attachments had ":" in their paths, and that character
doesn't work well on Windows, preventing the Git repo from even
being checked out.

This CL rewrites the names of those filenames to use '%3A' instead
of ':', and updates the scripts accordingly.

It also updates //.gitignore to not ignore the //scripts/feeds
directory, to make it a bit easier to remember to clear out the
feeds and start from scratch.

There are a couple other changes to files picked up as part
of the latest export as well.

Bug: 1268942
Change-Id: Idba426d58a6c5ee40edc4c8588ae15c5aa70387e