Make a windows config "work".
At least, it mostly seems to work. I hit a weird issue
with one PDF file throwing off the copy (possibly a
permissions thing), and Node seems to be incredibly
slow.

Also, the windows DEPS for node don't contain `npm`,
so `npmw` as written doesn't work. This isn't the
end of the world, as I've been meaning to replace
npmw and the run-scripts with something else anyway
to get rid of the dependency and make things a bit
clearer.

Bug: 1269902
Change-Id: Ieb18974f7b900431a40603fe87b317585242c875