Switch template engine to Nunjucks.
Eleventy uses the `Liquid` template engine by default, but
the other Eleventy-based sites that we maintain all use
Nunjucks, so this CL switches the engine to use Nunjucks
for compatibility with them.

Nunjucks seems to be a bit more prone to getting confused
than Liquid when template syntax like `{{`, `}}`, `{%` and
`%}` shows up unexpectedly in a file, so this CL updates
all of the site pages where Nunjucks was crashing the build
to work around this.

This CL also renames //site/_includes/page.html to
//site/_includes/page.njk to make the template format being
used more obvious.

Bug: 1269875
Change-Id: I23a793a5445f61e58f9b7b45f5a7869527ac4651