Add support for listing the subpages of a page.
This CL adds the ability for a page to include a list of all of
its subpages (in the sense that /developers/design-documents is
a subpage of /developers) automatically, rather than maintaining
a list manually.

To use this feature, add `{% subpages collections.all %}` to
your markdown. This results in a hierarchical tree of pages
being inserted into the HTML using <li> and <details> elements.

Bug: 1260453

Change-Id: I6613f01aaa7527eeea14d2d55bbab45339a2cf89