Update bot config to require successful builds before submitting.
Previously the try builder was configured to run as a Tricium
verifier only, so a failure did not keep a CL from submitting.

This change modifies the config to run as both a Tricium verifier
(to still get the preview comments posted to the bug) and as
a regular CQ job (so that we are required to have a passing build
before submitting the change.).

Change-Id: I0f7b6bd283fd2e03e101bc8816e4f9c6baf7ac22