Fix bugs introduced in refactoring of subpages.js.
When I refactored a bunch of the code in subpages.js to make
it more understandable, I also introduced some bugs that
caused pages to be duplicated in the list. This CL fixes the bugs.

This CL also changes the default state of the <details> elements
to `open` to match the behavior of the old site.

Bug: 1272014
Change-Id: I9382b0e7c3f72ff291d123654fa532da41fd4c9b