Clean up the CSS for the site.
This cleans up the CSS for the site by:

- Switching to using Sass variables for colors.
- Grouping the colors at the top of the stylesheet
  to be able to see the palette in one place.
- Sorting rules by name as much as possible.
- Adding some comments for sections.
- Stripping out rules that weren't being used.

There should be no particularly visible changes
to the site as a result. There were a few changes
where colors were tweaked slightly to reduce the
number of different colors, but you probably
wouldn't notice even if you were looking for it.

Change-Id: Ie9293222d068a4f53f74117ca4a6a4b48ec8e5a8