First pass at integrating Algolia search.
This doesn't actually work yet, but it's a skeleton.
For reasons that aren't clear to me yet, things are crashing
in the Algolia JS code.

This version disables CSP, so we don't want to actually land it.
Rather, we need to port the inline-script hashing code over from
DCC or some other example site.

Commit: false
Bug: 1260463
Change-Id: Idc5f9965da8e75925c58d454ca547a386a75a087