Generate firebase.json on the fly to support CSP properly.
In order to support CSP (Content Security Policy) properly
on a static website, we have to:

1) Move from using <script src=...> to inline script tags
   that dynamically add scripts to the page.
2) Compute the hashes of each script block at build time
   to add them to the allowed list of script blocks served by CSP.
3) Dynamically generate a firebase.json file (again at build time)
   that contains the allowed list of script blocks.

This CL does all of the above. If it detects during the build
that there are <script> tags anywhere on the site (which, at
the moment, will always be true due to the use of a dummy block
in the page template), it'll set the CSP 'script-src',
'unsafe-inline', and 'script-dynamic' directives in
the HTTP response. If there are no script tags, it'll send 'none'.

The implementation in this CL will serve the same header for
all resources on the site. Firebase appears to allow us to vary
the header by glob patterns, but we don't currently need that.
However, if we want to vary the scripts per-page and not bloat
every reply with unneeded hashes, we'll need to do something like
that.

Commit: false
Bug: 1260463
Change-Id: Idc5f9965da8e75925c58d454ca547a386a75a087