Fork docs/sync/ in docs/website/ (https://chromium.org)
We should keep these docs in a single place for simpler maintenance.
docs/website/ seems to be strictly better option, because besides
  * Appearing on code search
  * Being rendered by gittiles
like docs/sync/, it will also
  * Be rendered in chromium.org, which is often used by new
  contributors

For the moment we just copy the content to docs/website/ because the
directories are in different repositories. A second CL will follow
to delete docs/sync/.

Bug: 1006699
Change-Id: I55b6565c203686965e078bc42c63f918cd894436