Speed up downloading of the LOBs.
For some reason, trying to download all of the LOBs in the repo
one-at-a-time via download_from_google_storage is *very* slow (
it may take 30 min or more to fetch everything).

After experimenting with a few different approachs (a single tarball,
connection pooling, HTTPS vs. HTTP/2 vs. HTTP/3, Python vs. curl, etc.)
it looks like fetching everything in Python using urllib3 with
connection pools lets me fetch everything in ~30s over a good network
on a high-core machine.

Using a single tarball can get this down to ~6s, but that would
significantly complicate how we did things (do we need to upload
new tarballs every time a single LOB changes, etc.). It's not clear
that using HTTP/2 or HTTP/3 really helps, as the additional overhead
of the multiple objects may be server-side in GCS.

The fetch is incremental (will only download new things) and is a
no-op if everything is up-to-date, so hopefully this is an acceptable
compromise.

We should figure out why download_from_google_storage is so much
slower, but that can be done at a later date.

Bug: 1266070
Change-Id: Ib37353870f89eca43bdaf3547779643919b28850