Add missed "build" directories from export
The top-level /.gitignore file had entries like 'build' instead of
'/build', meaning that it would match *any* files or directories with those
names, rather than just the files in the top directories.

This CL fixes the .gitignore file and adds the two 'build' subdirs that
were missed in the original export.

Bug: 1295946
Change-Id: Ib7fd13a533fae4399f60001881135f10c5c68381