Replace October highlights and add November, January and February blink-interactions highlights
We are linking our highlights directly to Google Docs with Alphabet
access rather than using Markdown. We are moving away from the Markdown
because all of the existing ones have had their format interrupted,
and also we can write once and send it out as a nice formatted email
and we don't have to rewrite it again in Markdown.

Change-Id: I05469351f8546a6eed8903fe161d552c48403137