Remove inactive owners.
This is the result of running the remove_inactive_owners script,
removing owners that haven't reviewed or authored a CL in the last 12
months.
There may be cases where owners are actually active elsewhere in or
there is simply very low code change in a directory.
If a removed owner is actually still active, please comment.

https://chromium.googlesource.com/chromium/src/+/master/docs/code_reviews.md#expectations-of-owners

Bug: 877528
Change-Id: I75f484674eaf4fc6250596ed310b1c9e49dd94c2