[website] Update design-documents/profile-architecture
This page was written back in 2012, in the middle of the transition to
KeyedServices. It gave instructions to migrate code to the new
architecture, and generally contrasted the "old" (from 10 years ago!)
and "new" ways of doing things.

Update the article so it's more accurate to the current state of things.
Remove the section on migrating your code, since the migration was
completed a long time ago. Instead, give an example of how to add a new
KeyedService + factory in Chrome.

Also remove the section on the dependency visualizer, as it's not really
useful anymore (the dependency graph is way too large).

Old ToC:
(Intro)
- Design Goals
- BrowserContextKeyedServiceFactory
  - The Old Way: Profile interface and ProfileImpl
  - The New Way: BrowserContextKeyedServiceFactory
  - A Few Types of Factories
  - A Brief Interlude About Complexity
  - Dependency Management Overview
- The Five Minute Tutorial of How to Convert Your Code
- Debugging Tips
  - Using the dependency visualizer
  - Crashes at Shutdown

New ToC:
(Intro)
- Design Goals
- How to Add a New Service
  - KeyedService
  - BrowserContextKeyedServiceFactory
  - Use the Service
- A Few Types of Factories
- Two-phase Shutdown
- Dependency Management Overview
- Debugging Tips
  - Crashes at Shutdown

Bug: 1053445
Change-Id: If6af5e79649882d6287927b4b897e2cf3d7038ee