Trim away outdated parts of the SSL/TLS section
"Except on iOS" is no longer true when it comes to the net stack. It
is arguably true for web content because Chrome for iOS uses
WKWebView, but this document's about the net stack. Remove the
bit about certificate verification as that's in the process of moving,
and the bit about caching is long out of date.

Bug: 528999
Change-Id: If303d6193e6091aa7c3bada2e1c3030f8df643e5