Add a sitemap as /sitemap.xml.
Having a sitemap for the site should help Algolia (and Google)
find all of the pages on the site, since not every page is
reachable from the /chromium-projects page.

The sitemap file format is documented at:

  https://www.sitemaps.org/protocol.html

It is a very simple XML-based format.

In theory we could add a per-url <lastmod> field to the sitemap
to give crawlers a better hint at what's changing, but in order
to do that correctly we'd have to look it up via git, and I'm
not keen to do 1400+ git lookups in order to build the file,
so this CL leaves it out for now.  If it turns out that it
would be helpful, we can add something easily enough.

This CL also adds a /robots.txt file to tell crawlers about the
sitemap.

I used the following documentation to help with this CL:

https://developers.google.com/search/docs/crawling-indexing/robots/intro
https://developers.google.com/search/docs/crawling-indexing/sitemaps/overview

There's probably a better way of adding just Nunjucks-based
templating (i.e., files that don't use the default
/_includes/page.html) to this but I couldn't figure it
out in an hour or two of digging.

Bug: 1260455
Change-Id: I83d4eac60cde0d57b7896ce26dc5ddf5f17d7a5e