Update video linked from "What is Google Chrome OS?".
Video in old link is marked private. Bassed on my discovery of a
Korean subtitled version of the video
(https://www.youtube.com/watch?v=8cUGHLrrGO0), researching the studio
credited in the final seconds of the video, the video behind this link
(https://www.youtube.com/watch?v=LwVX42Amcak) is equivalent to the
original.

Change-Id: I28e0c524d6a379704b54981ba0a970a3a9cf60bd