Added ProfileKeyedServiceFactory (PKSF) information to the documentation
Enrich Profile-architecture documentation with the advantages and usage of `ProfileKeyedServiceFactory`.

Shifted the focus of the document to encourage the usage of PKSF over BCKSF when possible (which should be most of the usages).

The rest of the content is almost the same, except when mentioning the added value of PKSF over BCKSF; mainly over the control of the service creation per profile type.

Change-Id: Id176622b5ad5a13f76b909bf8ea07d4c3b9b2eb1
Bug: 1284664