Make temporary OT changes permanent.
The list of changes are as follow:

An OT can run for 6 milestones by default. If an extension is requested, it can only be extended for 3 milestones at a time and only if substantial progress is demonstrated towards shipping
(see the docs at chromium.org/blink/launching-features/#origin-trials for the full list).

Change-Id: Ia7853b49fbe9f4e408b7404d8f1602d90c414274