chromium-os: Add a new how-to on building and flashing custom firmware
There's a lot of great external documentation on how to put Matt
DeViller's TianoCore builds (for example) on a CrOS device.  However,
little documentation exists on how to build from source.  For those of
us who won't take some guy's binaries as an answer, I threw together a
little guide on how to do this.

Change-Id: I359175e27c81e24be18522538d37ffd3907fe695