Update windbg instructions for WinDbgX, !chkimg
Discussions of how to detect code corruption/patching led to several
useful suggestions of non-obvious windbg commands that can be used to
detect modified code bytes, and detect whether the modified-code-byte
detection was meaningful.

This change also adds references to the new WinDbgX, changes the
spelling of WinDBG to WinDbg (per WinDbg), and fixes/tidies the lm vm
command and explanation.

Change-Id: I889f53811cec707416340a04c694a0961cf1c713