Upgrade node from 14.15.4 to 16.13.0.
There are no code changes other than upgrading node and
no content changes, so the site content in //build should
be identical to before.

The binaries being used are the same as the binaries used
in chromium/src at this point in time. We should generally
try to track chromium/src's version.

Bug:1457689
Change-Id: I8bbdf64488439f821c9a6f1d6037250b1644d8eb