Update a few node packages to close some issues.
There are (at least) 4 projects out of date in the //node_modules
package. What I really need to do is go through and update everything
to be current, but in the interest of expediency, this is a CL
to just update the four I currently have bugs on.

The CL should cause no changes to the content.

Bug: b/286839583, b/286529310, b/286043254, b/285954099
Change-Id: Ibc7165e9fde3c446acc974e422d56f282b9aa820