website: Use host-relative URLs for several index pages
Using http://www.chromium.org/ URLs means that browsing a developer
instance of these pages can be confusing, as one ends up back on the
live site instead of the dev copy. Make things a bit more consistent,
since many index URLs already use host-relative paths.

Since it's not trivial to do automatic transformation for some of this,
I only bother with the first few levels of the tree directory tree, as
it's less likely devs are navigating the hierarchy the further down we
go.

Change-Id: I9e4749c14e5d8495aaf5808ec4dc5a17d5d4f78e