Replace the Privacy Whitepaper link with one to Privacy Tours
The Privacy Whitepaper is no longer being updated. Instead, the external
privacy documentation about Safe Browsing has been migrated to a new
format of a Privacy Tours article.

The new article does explain that additional data is only sent in case
of security incidents, so it is still relevant for this paragraph.

Bug: 1515893
Change-Id: I8f8b79a81fec94fd645792febd7b54222c1bbd82