Sync docs: Replace a reference to TypedUrlSpecifics
The TYPED_URLS data type has been deprecated for some time. While the
specifics proto still exists for now, it's best for our documentation
to not refer to deprecated stuff.
So this CL trivially replaces the reference with one to ReadingList
(kind of arbitrarily chosen, it's just a specifics proto that's neither
too simple nor too big/complicated).

Bug: 40059424
Change-Id: I32fa4d5efbf484321b64a7760be3b2fb89412a63