Upgrade node to 20.11.
This updates the binaries used to build build the website to
Node v20.11. This also switches how the binaries are fetched
to use the first-class support for GCS packages that has been
added to gclient.

The new versions of node (which are the same that Chromium uses)
do not contain an `npm` binary. I've modified the `npmw` wrapper
script to invoke node + `npm-run-all` directly to work around this.

This CL does not update any of the node packages themselves.
If/when we choose to do that, you'll have to supply your own version
of `npm` to do so (this is the same process they use in Chromium).