reorder text to better separate shell and trad flow
The old order of headers [1] suggested that chrome_sdk was applicable to
the entire simple chrome environment. But, it only is to be used in the
trad flow- this is better reflected in the new [2] order. I also moved
and emphasized a note that the steps under "Set up the ChromeOS device"
apply to both flows, but, the shell-less flow requires the full path.

[1]: http://screen/6LuCGedLp8degum
[2]: http://screen/7PSsuDBuJi5VzeD

Change-Id: I6f043c9b966bb9ae5c9509a8403800acff805080