Remove brucedawson@ from OWNERS and docs
I am leaving Google and won't be working on Chromium regularly so I need
to be removed from OWNERS files and other places. This OWNERS file has
enough coverage so no replacement is needed. There's no clear
replacement for the "need assistance" reference but I hope that the
documentation is sufficient.

Change-Id: I8fa5c87f9e6576e84337b12b436b1c451ed6014f