Standardize on "non-trivial" in becoming-a-committer.
We were using both "non-trivial" and "nontrivial" on the
becoming-a-committer page. Both appear to be accepted usage
and different sources recommend each. So it doesn't really
matter which we use, but we should probably at least try to
be consistent. It looks like we use "non-trivial" more often
than "trivial" elsewhere on the site, so I'll use that here.

Change-Id: I323120e74e07757d69a17c560e127cd3c890b5df