PRESUBMIT: check for /site/ prefixes in links
Sometimes people create local links with /site/ because the files live
under the site/ dir, and because those are the paths that gitiles uses.
So if developers spot check against gitiles, they'll think they need to
use /site/ in the link, otherwise it's broken.  But when the website is
generated, it uses site/ as its base path, and leading /site/ in the
link breaks things.  There are commits in this repo's history showing
people making this exact mistake.  So check for its usage.

Change-Id: I9bcb8eabc689d56956ae185e5aa99f3c3c85a41d