navbar: add a bookmarklet to help navigation
Add a code that helps people navigating to the same document that is
properly rendered.

The navbar says to visit www.chromium.org, but it doesn't navigate us to
the correspoding page. This change mitigates the issue.
It seems Gitiles dones't support javascript: scheme, so we let the
viewer bookmark the code by drag and drop and use it.

Bug: None
Change-Id: I54a06e1cabcb3d7af45cff3deef49457ebb3ab76